Nvidia-NIM/requirements.txt
openai
python-dotenv
langchain_nvidia_ai_endpoints
langchain_community
streamlit
pypdf